+++ keras_cv/models/__init__.py
@@ -12,6 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from keras_cv.models.backbones.convnext.convnext_backbone import ConvNeXtBase
+from keras_cv.models.backbones.convnext.convnext_backbone import ConvNeXtLarge
+from keras_cv.models.backbones.convnext.convnext_backbone import ConvNeXtSmall
+from keras_cv.models.backbones.convnext.convnext_backbone import ConvNeXtTiny
+from keras_cv.models.backbones.convnext.convnext_backbone import ConvNeXtXLarge
 from keras_cv.models.backbones.csp_darknet.csp_darknet_backbone import (
     CSPDarkNetBackbone,
 )
@@ -81,11 +86,6 @@ from keras_cv.models.convmixer import ConvMixer_768_32
 from keras_cv.models.convmixer import ConvMixer_1024_16
 from keras_cv.models.convmixer import ConvMixer_1536_20
 from keras_cv.models.convmixer import ConvMixer_1536_24
-from keras_cv.models.convnext import ConvNeXtBase
-from keras_cv.models.convnext import ConvNeXtLarge
-from keras_cv.models.convnext import ConvNeXtSmall
-from keras_cv.models.convnext import ConvNeXtTiny
-from keras_cv.models.convnext import ConvNeXtXLarge
 from keras_cv.models.darknet import DarkNet21
 from keras_cv.models.darknet import DarkNet53
 from keras_cv.models.densenet import DenseNet121
